+++ src/Order/Cart/ViewCartPage.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace SwipeStripe\Order\Cart;
+
+use SwipeStripe\Pages\RequiredSinglePage;
+
+/**
+ * Class ViewCartPage
+ * @package SwipeStripe\Order\Cart
+ */
+class ViewCartPage extends \Page
+{
+    use RequiredSinglePage;
+}
