--- src/Pages/ViewCartPage.php
@@ -1,13 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace SwipeStripe\Pages;
-
-/**
- * Class CartPage
- * @package SwipeStripe\Pages
- */
-class ViewCartPage extends \Page
-{
-    use RequiredSinglePage;
-}
