+++ tests/Pages/RequiredSinglePagesCreatedTest.php
@@ -5,9 +5,9 @@ namespace SwipeStripe\Tests\Pages;
 
 use SilverStripe\CMS\Model\SiteTree;
 use SilverStripe\Dev\SapphireTest;
+use SwipeStripe\Order\Cart\ViewCartPage;
 use SwipeStripe\Order\Checkout\CheckoutPage;
 use SwipeStripe\Pages\OrderConfirmationPage;
-use SwipeStripe\Pages\ViewCartPage;
 use SwipeStripe\Pages\ViewOrderPage;
 
 /**
